requirements.txt
pillow
tk
python-vlc